fix: break circular import between mcp/plan_tools and routes/api
mcp/plan_tools imported _build_coordinator_task from routes/api/_shared,
which triggered routes/api/__init__.py, which imported routes/api/plan,
which imported mcp/plan_tools — deadlocking the MCP server on startup.

Move all three _build_*_task helpers to services/task_builders (where
they architecturally belong — services, not route handlers). routes/api/_shared
now re-exports them for backward compatibility; mcp/plan_tools imports
directly from services/task_builders, cutting the routes/api dependency
entirely.
